+++ JustShare/bin/main/static/mapper/main-mapper.xml
@@ -3,25 +3,12 @@
 
 <mapper namespace="com.js.web.main.MainDAO">
 
-<select id="login" parameterType="Map" resultType="Map">
-		SELECT count(*) as count, mname, mid
-		FROM members
-		WHERE mid = #{mid } AND mpw = #{mpw}
+<select id="bannerlist" resultType="com.js.web.main.MainDTO"> 
+SELECT * FROM banner ORDER BY bano ASC 
 </select>
 
-<select id="findId" parameterType="Map" resultType="String">
-		SELECT MID 
-		FROM members 
-		WHERE mname = #{mname} AND mphone = #{mphone}
+<select id="cateList" resultType="Map">
+SELECT * FROM cate
 </select>
 
-<select id="findPwCheck" parameterType="Map" resultType="Map">
-		SELECT COUNT(*) AS count, memail FROM members 
-		WHERE mid = #{mid} AND mphone = #{mphone}
-</select>
-
-<update id="temporaryPw" parameterType="Map">
-		UPDATE members SET mpw = #{uuid}
-		WHERE memail = #{to}
-</update>
 </mapper>
