+++ JustShare/src/main/java/com/js/web/main/MainDAO.java
@@ -1,20 +1,16 @@
 package com.js.web.main;
 
+import java.util.List;
 import java.util.Map;
 
 import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
 public interface MainDAO {
-	Map<String, Object> login(Map<String, Object> map);
-
-	String findId(Map<String, Object> map);
-
-	Map<String, Object> findPW(Map<String, Object> map);
-
-	Map<String, Object> findPwCheck(Map<String, Object> map);
-
-	void temporaryPw(Map<String, Object> map);
 
 
+	List<MainDTO> bannerlist();
+
+	List<Map<String, Object>> cateList();
+	
 }
