+++ JustShare/src/main/java/com/js/web/main/MainDTO.java
@@ -1,0 +1,10 @@
+package com.js.web.main;
+
+import lombok.Data;
+
+@Data
+public class MainDTO {
+private int bano;
+private String batext,baimg; 	
+	
+}
