+++ JustShare/src/main/java/com/js/web/main/MainService.java
@@ -1,5 +1,6 @@
 package com.js.web.main;
 
+import java.util.List;
 import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,24 +12,18 @@ public class MainService {
 	@Autowired 
 	private MainDAO mainDAO;
 
-	public Map<String, Object> login(Map<String, Object> map) {
+
+
+	public List<MainDTO> bannerlist() {
 		
-		return mainDAO.login(map);
+		return mainDAO.bannerlist();
 	}
 
-	public String findId(Map<String, Object> map) {
-		
-		return mainDAO.findId(map);
-	}
 
-	public Map<String, Object> findPwCheck(Map<String, Object> map) {
-		
-		return mainDAO.findPwCheck(map);
-	}
 
-	public void temporaryPw(Map<String, Object> map) {
-			mainDAO.temporaryPw(map);
-		
+	public List<Map<String, Object>> cateList() {
+	
+		return mainDAO.cateList();
 	}
