+++ JustShare/src/main/resources/static/mapper/login-mapper.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="com.js.web.login.LoginDAO">
+
+<select id="login" parameterType="Map" resultType="Map">
+		SELECT count(*) as count, mname, mid
+		FROM members
+		WHERE mid = #{mid } AND mpw = #{mpw}
+</select>
+
+<select id="findId" parameterType="Map" resultType="String">
+		SELECT MID 
+		FROM members 
+		WHERE mname = #{mname} AND mphone = #{mphone}
+</select>
+
+<select id="findPwCheck" parameterType="Map" resultType="Map">
+		SELECT COUNT(*) AS count, memail FROM members 
+		WHERE mid = #{mid} AND mphone = #{mphone}
+</select>
+
+<update id="temporaryPw" parameterType="Map">
+		UPDATE members SET mpw = #{uuid}
+		WHERE memail = #{to}
+</update>
+
+
+
+</mapper>
